Fix Java 9 warnings
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/batik/trunk@1802297 13f79535-47bb-0310-9956-ffa450edef68

+++ batik-bridge/src/main/java/org/apache/batik/bridge/StrokingTextPainter.java
@@ -441,7 +441,7 @@ public class StrokingTextPainter extends BasicTextPainter {
             float fontSize = 12;
             Float fsFloat = (Float)aci.getAttribute(TextAttribute.SIZE);
             if (fsFloat != null)
-                fontSize = fsFloat.floatValue();
+                fontSize = fsFloat;
 
             // if could not resolve at least one of the fontFamilies
             // then use the default font
@@ -723,23 +723,23 @@ public class StrokingTextPainter extends BasicTextPainter {
             if (layout.isVertical()) {
                 if (lengthAdj == ADJUST_SPACING) {
                     yScale = (float)
-                        ((length.floatValue() - lastH) /
+                        ((length - lastH) /
                          (advance.getY() - lastMetrics.getVerticalAdvance()));
                 } else {
                     double adv =(advance.getY() + lastH -
                                  lastMetrics.getVerticalAdvance());
-                    yScale = (float)(length.floatValue()/adv);
+                    yScale = (float)(length /adv);
                 }
                 visualAdvance = new Point2D.Float(0, length.floatValue());
             } else {
                 if (lengthAdj == ADJUST_SPACING) {
                     xScale = (float)
-                        ((length.floatValue() - lastW) /
+                        ((length - lastW) /
                          (advance.getX() - lastMetrics.getHorizontalAdvance()));
                 } else {
                     double adv = (advance.getX() + lastW -
                                   lastMetrics.getHorizontalAdvance());
-                    xScale = (float)(length.floatValue()/adv);
+                    xScale = (float)(length /adv);
                 }
                 visualAdvance = new Point2D.Float(length.floatValue(), 0);
             }
@@ -791,12 +791,12 @@ public class StrokingTextPainter extends BasicTextPainter {
         // Of course X and Y override that, but they don't apply for
         // text on a path.
         if ((runX != null) && (!runX.isNaN())) {
-            absX = runX.floatValue();
+            absX = runX;
             tpShiftX = absX;
         }
 
         if ((runY != null) && (!runY.isNaN())) {
-            absY = runY.floatValue();
+            absY = runY;
             tpShiftY = absY;
         }
 
@@ -821,12 +821,12 @@ public class StrokingTextPainter extends BasicTextPainter {
             if (vertical) {
                 runX = (Float) runaci.getAttribute(XPOS);
                 if ((runX != null) && (!runX.isNaN())) {
-                    absX = runX.floatValue();
+                    absX = runX;
                 }
             } else {
                 runY = (Float) runaci.getAttribute(YPOS);
                 if ((runY != null) && (!runY.isNaN())) {
-                    absY = runY.floatValue();
+                    absY = runY;
                 }
             }
 
@@ -1677,7 +1677,7 @@ public class StrokingTextPainter extends BasicTextPainter {
 
             Integer level = (Integer) aci.getAttribute(BIDI_LEVEL);
             if (level != null) {
-                this.level = level.intValue();
+                this.level = level;
             } else {
                 this.level = -1;
             }
